+++ NEWS.md
@@ -8,7 +8,7 @@
 
 * Improved performance for `merge_df()`. Furthermore, `add_rows()` was added as alias for `merge_df()`.
 * `merge_df()` resp. `add_rows()` now create a unique `id`-name instead of dropping the ID-variable, in case `id` has the same name of any existing variables in the provides data frames.
-* Improved performance for `descr()`.
+* Improved performance for `descr()` and minor changes to the output.
 
 ## Support for `mids`-objects (package _mice_)